+++ crash_handler/CrashReportApp.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <wxx_wincore.h>
+#include <crash_handler_stub.h>
 
 class CommandLineInfo;
 
@@ -12,8 +13,11 @@ public:
 
 private:
     void PrepareReport(const CommandLineInfo& cmd_line_info);
+    void ArchiveReport(const char* crash_dump_filename, const char* exc_info_filename);
     void SendReport();
     int Message(HWND hwnd, const char *pszText, const char *pszTitle, int Flags);
+
+    CrashHandlerConfig m_config;
 };
 
 inline CrashReportApp* GetCrashReportApp()